Actually move everything into cache_utils.py
