+++ lib/streamlit/caching/cache_utils.py
@@ -1,0 +1,196 @@
+# Copyright 2018-2021 Streamlit Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Shared caching-related utilities."""
+
+import hashlib
+import inspect
+import threading
+import types
+from typing import List, Tuple, Optional, Any, Union
+
+from streamlit import type_util, util
+from streamlit.errors import StreamlitAPIException
+from streamlit.logger import get_logger
+
+from .hashing import HashReason, update_hash, UnhashableTypeError
+
+_LOGGER = get_logger(__name__)
+
+
+def make_value_key(func: types.FunctionType, *args, **kwargs) -> str:
+    """Create the key for a value within a cache.
+
+    This key is generated from the function's arguments. All arguments
+    will be hashed, except for those named with a leading "_".
+
+    Raises
+    ------
+    StreamlitAPIException
+        Raised (with a nicely-formatted explanation message) if we encounter
+        an un-hashable arg.
+    """
+
+    # Create a (name, value) list of all *args and **kwargs passed to the
+    # function.
+    arg_pairs: List[Tuple[Optional[str], Any]] = []
+    for arg_idx in range(len(args)):
+        arg_name = _get_positional_arg_name(func, arg_idx)
+        arg_pairs.append((arg_name, args[arg_idx]))
+
+    for kw_name, kw_val in kwargs.items():
+        # **kwargs ordering is preserved, per PEP 468
+        # https://www.python.org/dev/peps/pep-0468/, so this iteration is
+        # deterministic.
+        arg_pairs.append((kw_name, kw_val))
+
+    # Create the hash from each arg value, except for those args whose name
+    # starts with "_". (Underscore-prefixed args are deliberately excluded from
+    # hashing.)
+    args_hasher = hashlib.new("md5")
+    for arg_name, arg_value in arg_pairs:
+        if arg_name is not None and arg_name.startswith("_"):
+            _LOGGER.debug("Not hashing %s because it starts with _", arg_name)
+            continue
+
+        try:
+            update_hash(
+                (arg_name, arg_value),
+                hasher=args_hasher,
+                hash_reason=HashReason.CACHING_FUNC_ARGS,
+                hash_source=func,
+            )
+        except UnhashableTypeError as exc:
+            raise StreamlitAPIException(
+                _get_unhashable_arg_message(func, arg_name, arg_value)
+            ) from exc
+
+    value_key = args_hasher.hexdigest()
+    _LOGGER.debug("Cache key: %s", value_key)
+
+    return value_key
+
+
+def make_function_key(func: types.FunctionType) -> str:
+    """Create the unique key for a function's cache.
+
+    A naive implementation would involve simply creating the cache object
+    right in the wrapper, which in a normal Python script would be executed
+    only once. But in Streamlit, we reload all modules related to a user's
+    app when the app is re-run, which means that - among other things - all
+    function decorators in the app will be re-run, and so any decorator-local
+    objects will be recreated.
+
+    Furthermore, our caches can be destroyed and recreated (in response to
+    cache clearing, for example), which means that retrieving the function's
+    cache in the decorator (so that the wrapped function can save a lookup)
+    is incorrect: the cache itself may be recreated between
+    decorator-evaluation time and decorated-function-execution time. So we
+    must retrieve the cache object *and* perform the cached-value lookup
+    inside the decorated function.
+    """
+    func_hasher = hashlib.new("md5")
+
+    # Include the function's __module__ and __qualname__ strings in the hash.
+    # This means that two identical functions in different modules
+    # will not share a hash; it also means that two identical *nested*
+    # functions in the same module will not share a hash.
+    update_hash(
+        (func.__module__, func.__qualname__),
+        hasher=func_hasher,
+        hash_reason=HashReason.CACHING_FUNC_BODY,
+        hash_source=func,
+    )
+
+    # Include the function's source code in its hash. If the source code can't
+    # be retrieved, fall back to the function's bytecode instead.
+    source_code: Union[str, types.CodeType]
+    try:
+        source_code = inspect.getsource(func)
+    except OSError as e:
+        _LOGGER.debug(
+            "Failed to retrieve function's source code when building its key; falling back to bytecode. err={0}",
+            e,
+        )
+        source_code = func.__code__
+
+    update_hash(
+        source_code,
+        hasher=func_hasher,
+        hash_reason=HashReason.CACHING_FUNC_BODY,
+        hash_source=func,
+    )
+
+    cache_key = func_hasher.hexdigest()
+    return cache_key
+
+
+def _get_positional_arg_name(func: types.FunctionType, arg_index: int) -> Optional[str]:
+    """Return the name of a function's positional argument.
+
+    If arg_index is out of range, or refers to a parameter that is not a
+    named positional argument (e.g. an *args, **kwargs, or keyword-only param),
+    return None instead.
+    """
+    if arg_index < 0:
+        return None
+
+    params: List[inspect.Parameter] = list(inspect.signature(func).parameters.values())
+    if arg_index >= len(params):
+        return None
+
+    if params[arg_index].kind in (
+        inspect.Parameter.POSITIONAL_OR_KEYWORD,
+        inspect.Parameter.POSITIONAL_ONLY,
+    ):
+        return params[arg_index].name
+
+    return None
+
+
+def _get_unhashable_arg_message(
+    func: types.FunctionType, arg_name: Optional[str], arg_value: Any
+) -> str:
+    arg_name_str = arg_name if arg_name is not None else "(unnamed)"
+    arg_type = type_util.get_fqn_type(arg_value)
+    func_name = func.__name__
+    arg_replacement_name = f"_{arg_name}" if arg_name is not None else "_arg"
+
+    return (
+        f"""
+Cannot hash argument '{arg_name_str}' (of type `{arg_type}`) in '{func_name}'.
+
+To address this, you can tell @st.memo not to hash this argument by adding a
+leading underscore to the argument's name in the function signature:
+
+```
+@st.memo
+def {func_name}({arg_replacement_name}, ...):
+    ...
+```
+        """
+    ).strip("\n")
+
+
+class ThreadLocalCacheInfo(threading.local):
+    """A thread-local counter that's incremented when we enter a cache function
+    and decremented when we exit.
+    """
+
+    def __init__(self):
+        self.cached_func_stack: List[types.FunctionType] = []
+        self.suppress_st_function_warning = 0
+
+    def __repr__(self) -> str:
+        return util.repr_(self)
